Restructure generator into modular architecture
- Extract common utilities into separate common.go file
- Add dependency sorting with topological sort algorithm
- Split generator into single-file and multi-file implementations
- Move template functions and type conversion logic to shared utilities
- Implement proper dependency resolution for TypeScript generation
- Maintain backward compatibility with existing generator interface
